refactor: dialog spacing, toast after removal, remove unncessary css class (#13)
