+++ packages/ui/public/locales/en/translation.json
@@ -9,6 +9,7 @@
   "Accounts": "",
   "Address copied!": "",
   "Address format": "",
+  "All authorized dapps have been removed": "",
   "An application, self-identifying as request app name is requesting access your wallet from origin.": "An application, self-identifying as <strong>{{appName}}</strong> is requesting access your wallet from <strong>{{origin}}</strong>.",
   "Approve Transaction": "",
   "Are you sure to remove all dapps access to your wallet?": "",
@@ -103,7 +104,7 @@
   "Scan QR Code": "",
   "Scan QR Code from Coong Wallet on a different device": "",
   "Search by name": "",
-  "Search by name or url": "",
+  "Search by name or URL": "",
   "Secret Recovery Phrase": "",
   "Secret recovery phrase": "",
   "Secret recovery phrase or password are missing": "",
