+++ packages/ui/src/components/shared/settings/ManageDappAccessDialog/DappAccessItem.tsx
@@ -10,9 +10,9 @@ export default function DappAccessItem({ appInfo }: DappAccessItemProps): JSX.El
   const appId = id || walletState.extractAppId(url);
 
   return (
-    <div className='border border-black/10 dark:border-white/15 py-2 px-4 rounded flex flex-row items-center gap-4 bg-zinc-50 dark:bg-black/15'>
+    <div className='border border-black/10 dark:border-white/15 py-2 px-4 rounded flex items-center gap-4 bg-zinc-50 dark:bg-black/15'>
       <div className='flex-grow'>
-        <div className='flex flex-row items-center gap-4'>
+        <div className='flex items-center gap-4'>
           <img src={`https://icon.horse/icon/${appId}`} alt={`${name} icon`} width='18' />
           <div className='flex flex-col'>
             <div className='font-semibold'>{name}</div>
