+++ packages/ui/src/hooks/wallet/useAuthorizedApps.ts
